// src/models/enum-models/IndustryEnum.ts
export enum IndustryEnum {
  ENERGY_UTILITIES = 'Energy & Utilities',
  MANUFACTURING_INDUSTRIAL = 'Manufacturing & Industrial',
  TECHNOLOGY_TELECOMMUNICATIONS = 'Technology & Telecommunications',
  HEALTHCARE_PHARMACEUTICALS = 'Healthcare & Pharmaceuticals',
  FINANCIAL_SERVICES = 'Financial Services',
  RETAIL_CONSUMER_GOODS = 'Retail & Consumer Goods',
  TRANSPORTATION_LOGISTICS = 'Transportation & Logistics',
  REAL_ESTATE_CONSTRUCTION = 'Real Estate & Construction',
  MEDIA_ENTERTAINMENT = 'Media & Entertainment',
  AGRICULTURE_FOOD_INDUSTRY = 'Agriculture & Food Industry',
  OTHER = 'Other',
}